// lib/relay/__generated__/profileLikesListFragment.graphql.ts
export type BibleBook =
  | "ACTS"
  | "AMOS"
  | "COLOSSIANS"
  | "DANIEL"
  | "DEUTERONOMY"
  | "ECCLESIASTES"
  | "EPHESIANS"
  | "ESTHER"
  | "EXODUS"
  | "EZEKIEL"
  | "EZRA"
  | "FIRST_CHRONICLES"
  | "FIRST_CORINTHIANS"
  | "FIRST_JOHN"
  | "FIRST_KINGS"
  | "FIRST_PETER"
  | "FIRST_SAMUEL"
  | "FIRST_THESSALONIANS"
  | "FIRST_TIMOTHY"
  | "GALATIANS"
  | "GENESIS"
  | "HABAKKUK"
  | "HAGGAI"
  | "HEBREWS"
  | "HOSEA"
  | "ISAIAH"
  | "JAMES"
  | "JEREMIAH"
  | "JOB"
  | "JOEL"
  | "JOHN"
  | "JONAH"
  | "JOSHUA"
  | "JUDE"
  | "JUDGES"
  | "LAMENTATIONS"
  | "LEVITICUS"
  | "LUKE"
  | "MALACHI"
  | "MARK"
  | "MATTHEW"
  | "MICAH"
  | "NAHUM"
  | "NEHEMIAH"
  | "NUMBERS"
  | "OBADIAH"
  | "PHILEMON"
  | "PHILIPPIANS"
  | "PROVERBS"
  | "PSALMS"
  | "REVELATION"
  | "ROMANS"
  | "RUTH"
  | "SECOND_CHRONICLES"
  | "SECOND_CORINTHIANS"
  | "SECOND_JOHN"
  | "SECOND_KINGS"
  | "SECOND_PETER"
  | "SECOND_SAMUEL"
  | "SECOND_THESSALONIANS"
  | "SECOND_TIMOTHY"
  | "SONG_OF_SONGS"
  | "THIRD_JOHN"
  | "TITUS"
  | "ZECHARIAH"
  | "ZEPHANIAH"
  | "%future added value";
export type BibleTranslation = "ASV" | "KJV" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type profileLikesListFragment$data = {
  readonly id: string;
  readonly likedBibleVersePosts: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly childPostsCount: number;
        readonly content: string | null | undefined;
        readonly createdAt: any;
        readonly id: string;
        readonly images: ReadonlyArray<{
          readonly height: number | null | undefined;
          readonly url: string | null | undefined;
          readonly width: number | null | undefined;
        }>;
        readonly likedAt: any | null | undefined;
        readonly likesCount: number;
        readonly poll:
          | {
              readonly deadline: any | null | undefined;
              readonly id: string;
              readonly isExpired: boolean | null | undefined;
              readonly options:
                | ReadonlyArray<{
                    readonly id: string;
                    readonly text: string | null | undefined;
                    readonly voteCount: number | null | undefined;
                    readonly votePercentage: number | null | undefined;
                  }>
                | null
                | undefined;
              readonly totalVotes: number | null | undefined;
              readonly userVote:
                | {
                    readonly id: string | null | undefined;
                    readonly text: string | null | undefined;
                  }
                | null
                | undefined;
            }
          | null
          | undefined;
        readonly user: {
          readonly id: string;
          readonly image:
            | {
                readonly url: string | null | undefined;
              }
            | null
            | undefined;
          readonly name: string | null | undefined;
          readonly username: string | null | undefined;
        };
        readonly verse:
          | {
              readonly book: BibleBook;
              readonly chapter: number;
              readonly id: string;
              readonly translation: BibleTranslation;
              readonly verse: number;
            }
          | null
          | undefined;
      };
    }>;
  };
  readonly " $fragmentType": "profileLikesListFragment";
};
export type profileLikesListFragment$key = {
  readonly " $data"?: profileLikesListFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"profileLikesListFragment">;
};